fix update_pkt_len (int -> uint16)

+++ bpf/srv6_tracer.c
@@ -280,7 +280,7 @@ static __always_inline struct sr6_pktid_tlv *get_pktidtlv(struct ipv6_sr_hdr *sr
   return NULL;
 }
 
-static __always_inline bool update_pkt_len(struct ipv6hdr *ipv6, struct ipv6_sr_hdr *srh, int size)
+static __always_inline bool update_pkt_len(struct ipv6hdr *ipv6, struct ipv6_sr_hdr *srh, uint16_t size)
 {
   ipv6->payload_len += htons(size);
   srh->hdrlen += size / 8;
